dodałem walidację formularza do tworzenia mezocyklu i jego edycji pod kątem więzów unikalności w obrębie BD

+++ clubs/views.py
@@ -796,6 +796,7 @@ def mezocycles(request,team_id):
     return render(request,'clubs/mezocycles.html',{'teams':teams,'usersClubs':usersClubs, 'team':team, 'mezocycles':mezocycles,'implemented_mezocycles':implemented_mezocycles})
 
 def create_mezocycle(request, team_id):
+    print(request.POST)
 
     class Training_in_week:
         def __init__(self, week, training):
@@ -805,7 +806,7 @@ def create_mezocycle(request, team_id):
     usersClubs, teams = get_data_for_menu(request)
     team = get_object_or_404(Team,pk = team_id)
     user = request.user
-    mezocycle_form = MezocycleForm(request.POST or None)
+    mezocycle_form = MezocycleForm(request.POST or None,team=team)
     weeks = 0
     trainings_per_week = 0
     mezocycle_set = False
@@ -824,12 +825,10 @@ def create_mezocycle(request, team_id):
                 for_t = range(1, trainings_per_week + 1)
                 for_wt = list(product(for_w, for_t))
                 for w,t in for_wt:
-                    form = Training_in_mezocycleForm(request.POST or None,initial={'topic':'ee','duration':3}, prefix=(str(w)+str(t)))
+                    form = Training_in_mezocycleForm(request.POST or None,initial={'topic':'ee','duration':3}, prefix=(str(w)+'_'+str(t)))
                     wt = Training_in_week(w,t)
                     forms_list.append((wt,form))
-                
-                
-
+                    
         if 'mezocycle_details' in request.POST:
             if mezocycle_form.is_valid():
                 weeks = mezocycle_form.cleaned_data.get('weeks')
@@ -839,7 +838,7 @@ def create_mezocycle(request, team_id):
                 for_t = range(1, trainings_per_week + 1)
                 for_wt = list(product(for_w, for_t))
                 for w,t in for_wt:
-                    form = Training_in_mezocycleForm(request.POST or None,initial={'topic':'ee','duration':3}, prefix=(str(w)+str(t)))
+                    form = Training_in_mezocycleForm(request.POST or None,initial={'topic':'ee','duration':3}, prefix=(str(w)+'_'+str(t)))
                     wt = Training_in_week(w,t)
                     forms_list.append((wt,form))
 
@@ -862,13 +861,13 @@ def create_mezocycle(request, team_id):
 
                     for w,t in for_wt:
                         training_in_mezocycle = Training_in_mezocycle()
-                        training_in_mezocycle.topic = request.POST.get(str(w)+str(t)+"-topic", "")
-                        training_in_mezocycle.week_number = request.POST.get(str(w)+str(t)+"-week_number", "")
-                        training_in_mezocycle.training_number = request.POST.get(str(w)+str(t)+"-training_number", "")
-                        training_in_mezocycle.goals = request.POST.get(str(w)+str(t)+"-goals", "")
-                        training_in_mezocycle.rules = request.POST.get(str(w)+str(t)+"-rules", "")
-                        training_in_mezocycle.actions = request.POST.get(str(w)+str(t)+"-actions", "")
-                        training_in_mezocycle.duration = request.POST.get(str(w)+str(t)+"-duration", "")
+                        training_in_mezocycle.topic = request.POST.get(str(w)+'_'+str(t)+"-topic", "")
+                        training_in_mezocycle.week_number = request.POST.get(str(w)+'_'+str(t)+"-week_number", "")
+                        training_in_mezocycle.training_number = request.POST.get(str(w)+'_'+str(t)+"-training_number", "")
+                        training_in_mezocycle.goals = request.POST.get(str(w)+'_'+str(t)+"-goals", "")
+                        training_in_mezocycle.rules = request.POST.get(str(w)+'_'+str(t)+"-rules", "")
+                        training_in_mezocycle.actions = request.POST.get(str(w)+'_'+str(t)+"-actions", "")
+                        training_in_mezocycle.duration = request.POST.get(str(w)+'_'+str(t)+"-duration", "")
                         training_in_mezocycle.mezocycle = mezocycle
                         training_in_mezocycle.save()
                     return redirect(mezocycles,team.id)
@@ -878,7 +877,7 @@ def create_mezocycle(request, team_id):
         for_wt = list(product(for_w, for_t))
         print(request.POST)
         for w,t in for_wt:
-            form = Training_in_mezocycleForm(request.POST or None,initial={'topic':'ee','duration':3}, prefix=(str(w)+str(t)))
+            form = Training_in_mezocycleForm(request.POST or None,initial={'topic':'ee','duration':3}, prefix=(str(w)+'_'+str(t)))
             wt = Training_in_week(w,t)
             forms_list.append((wt,form))
 
@@ -889,33 +888,16 @@ def create_mezocycle(request, team_id):
 
 
 def edit_mezocycle(request, mezocycle_id):
+    print(request.POST)
     usersClubs, teams = get_data_for_menu(request)
     mezocycle = get_object_or_404(Mezocycle,pk = mezocycle_id)
     trainings = Training_in_mezocycle.objects.filter(mezocycle = mezocycle)
-    mezocycle_form = MezocycleForm(request.POST or None, instance=mezocycle)
+    mezocycle_form = MezocycleForm(request.POST or None, instance=mezocycle,team=mezocycle.team)
+    for_w = []
+    for_t = []
+    forms_list=[]
     if request.method == 'POST':
         if 'mezocycle' in request.POST:
-            if mezocycle_form.is_valid():
-                mezocycle = mezocycle_form.save(commit=False)
-                forms_list = []
-                for_w = range(1, mezocycle.weeks + 1)
-                for_t = range(1, mezocycle.trainings_per_week + 1)
-                weeks = trainings.order_by('-week_number').first()
-                tranings_per_week = trainings.order_by('-training_number').first()
-                for w in for_w:
-                    for t in for_t:
-                        for training in trainings:
-                            if training.week_number == w and training.training_number == t:
-                                form = Training_in_mezocycleForm(request.POST or None, instance=training, prefix=(str(training.week_number)+str(training.training_number)))
-                                form.week_number = training.week_number
-                                form.training_number =  training.training_number
-                                forms_list.append(form)
-                        if t > tranings_per_week.training_number or w > weeks.week_number:
-                            form = Training_in_mezocycleForm(request.POST or None, initial={'topic':None,'duration':None}, prefix=(str(w)+str(t)))
-                            form.week_number = w
-                            form.training_number = t
-                            forms_list.append(form)
-        if 'mezocycle_details' in request.POST:
             if mezocycle_form.is_valid():
                 mezocycle = mezocycle_form.save(commit=False)
             forms_list = []
@@ -927,35 +909,63 @@ def edit_mezocycle(request, mezocycle_id):
                 for t in for_t:
                     for training in trainings:
                         if training.week_number == w and training.training_number == t:
-                            form = Training_in_mezocycleForm(request.POST or None, instance=training, prefix=(str(training.week_number)+str(training.training_number)))
+                            form = Training_in_mezocycleForm(request.POST or None, instance=training, prefix=(str(training.week_number)+'_'+str(training.training_number)))
                             form.week_number = training.week_number
                             form.training_number =  training.training_number
                             forms_list.append(form)
                     if t > tranings_per_week.training_number or w > weeks.week_number:
-                        training_in_mezocycle = Training_in_mezocycle()
-                        training_in_mezocycle.topic = request.POST.get(str(w)+str(t)+"-topic", "")
-                        training_in_mezocycle.week_number = w
-                        training_in_mezocycle.training_number = t
-                        training_in_mezocycle.goals = request.POST.get(str(w)+str(t)+"-goals", "")
-                        training_in_mezocycle.rules = request.POST.get(str(w)+str(t)+"-rules", "")
-                        training_in_mezocycle.actions = request.POST.get(str(w)+str(t)+"-actions", "")
-                        training_in_mezocycle.duration = request.POST.get(str(w)+str(t)+"-duration", "")
-                        training_in_mezocycle.mezocycle = mezocycle
-                        form = Training_in_mezocycleForm(request.POST or None, instance=training_in_mezocycle, prefix=(str(w)+str(t)))
-                        form.week_number = training.week_number
-                        form.training_number =  training.training_number
+                        form = Training_in_mezocycleForm(request.POST or None, initial={'topic':None,'duration':None}, prefix=(str(w)+'_'+str(t)))
+                        form.week_number = w
+                        form.training_number = t
                         forms_list.append(form)
-            for form in forms_list:
-                if form.is_valid():
-                    form.save()
-                else:
-                    print('błąd walidacji')
-            mezocycle.save()
-            for training in trainings:
-                if training.week_number not in for_w or training.training_number not in for_t:
-                    training.delete()
-            return redirect(mezocycles,mezocycle.team.id)
-        
+        if 'mezocycle_details' in request.POST:
+            if mezocycle_form.is_valid():
+                mezocycle = mezocycle_form.save(commit=False)
+
+                forms_list = []
+                for_w = range(1, mezocycle.weeks + 1)
+                for_t = range(1, mezocycle.trainings_per_week + 1)
+                weeks = trainings.order_by('-week_number').first()
+                tranings_per_week = trainings.order_by('-training_number').first()
+                for w in for_w:
+                    for t in for_t:
+                        for training in trainings:
+                            if training.week_number == w and training.training_number == t:
+                                form = Training_in_mezocycleForm(request.POST or None, instance=training, prefix=(str(training.week_number)+'_'+str(training.training_number)))
+                                form.week_number = training.week_number
+                                form.training_number =  training.training_number
+                                forms_list.append(form)
+                        if t > tranings_per_week.training_number or w > weeks.week_number:
+                            training_in_mezocycle = Training_in_mezocycle()
+                            training_in_mezocycle.topic = request.POST.get(str(w)+'_'+str(t)+"-topic", "")
+                            training_in_mezocycle.week_number = w
+                            training_in_mezocycle.training_number = t
+                            training_in_mezocycle.goals = request.POST.get(str(w)+'_'+str(t)+"-goals", "")
+                            training_in_mezocycle.rules = request.POST.get(str(w)+'_'+str(t)+"-rules", "")
+                            training_in_mezocycle.actions = request.POST.get(str(w)+'_'+str(t)+"-actions", "")
+                            training_in_mezocycle.duration = request.POST.get(str(w)+'_'+str(t)+"-duration", "")
+                            training_in_mezocycle.mezocycle = mezocycle
+                            form = Training_in_mezocycleForm(request.POST or None, instance=training_in_mezocycle, prefix=(str(w)+'_'+str(t)))
+                            form.week_number = w
+                            form.training_number = t
+                            forms_list.append(form)
+
+                all_forms_valid = True
+                for form in forms_list:
+                    if not form.is_valid():
+                        print("brak walidacji")
+                        all_forms_valid=False
+
+                if all_forms_valid:
+                    for form in forms_list:
+                        if form.is_valid():
+                            form.save()
+                    mezocycle.save()
+                    for training in trainings:
+                        if training.week_number not in for_w or training.training_number not in for_t:
+                            training.delete()
+                    return redirect(mezocycles,mezocycle.team.id)
+    
         if 'mezocycle_copy' in request.POST:
             if mezocycle_form.is_valid():
                 print('mezocykl form valid')
@@ -969,45 +979,54 @@ def edit_mezocycle(request, mezocycle_id):
                 for t in for_t:
                     for training in trainings:
                         if training.week_number == w and training.training_number == t:
-                            form = Training_in_mezocycleForm(request.POST or None, instance=training, prefix=(str(training.week_number)+str(training.training_number)))
+                            form = Training_in_mezocycleForm(request.POST or None, instance=training, prefix=(str(training.week_number)+'_'+str(training.training_number)))
                             form.week_number = training.week_number
                             form.training_number =  training.training_number
                             forms_list.append(form)
                     if t > tranings_per_week.training_number or w > weeks.week_number:
                         training_in_mezocycle = Training_in_mezocycle()
-                        training_in_mezocycle.topic = request.POST.get(str(w)+str(t)+"-topic", "")
+                        training_in_mezocycle.topic = request.POST.get(str(w)+'_'+str(t)+"-topic", "")
                         training_in_mezocycle.week_number = w
                         training_in_mezocycle.training_number = t
-                        training_in_mezocycle.goals = request.POST.get(str(w)+str(t)+"-goals", "")
-                        training_in_mezocycle.rules = request.POST.get(str(w)+str(t)+"-rules", "")
-                        training_in_mezocycle.actions = request.POST.get(str(w)+str(t)+"-actions", "")
-                        training_in_mezocycle.duration = request.POST.get(str(w)+str(t)+"-duration", "")
+                        training_in_mezocycle.goals = request.POST.get(str(w)+'_'+str(t)+"-goals", "")
+                        training_in_mezocycle.rules = request.POST.get(str(w)+'_'+str(t)+"-rules", "")
+                        training_in_mezocycle.actions = request.POST.get(str(w)+'_'+str(t)+"-actions", "")
+                        training_in_mezocycle.duration = request.POST.get(str(w)+'_'+str(t)+"-duration", "")
                         training_in_mezocycle.mezocycle = mezocycle
-                        form = Training_in_mezocycleForm(request.POST or None, instance=training_in_mezocycle, prefix=(str(w)+str(t)))
-                        form.week_number = training.week_number
-                        form.training_number =  training.training_number
+                        form = Training_in_mezocycleForm(request.POST or None, instance=training_in_mezocycle, prefix=(str(w)+'_'+str(t)))
+                        form.week_number = w
+                        form.training_number =  t
                         forms_list.append(form)
             
-            new_mezocycle = Mezocycle(name=mezocycle.name, user=mezocycle.user, team=mezocycle.team, weeks=mezocycle.weeks, trainings_per_week=mezocycle.trainings_per_week)
-            new_mezocycle.save()
-            print(mezocycle,'|',new_mezocycle)
-            for form in forms_list:
-                if form.is_valid():
-                    training = form.save(commit=False)
-                    new_training = Training_in_mezocycle()
-                    new_training.mezocycle = new_mezocycle
-                    new_training.topic = training.topic
-                    new_training.week_number = training.week_number
-                    new_training.training_number = training.training_number
-                    new_training.goals = training.goals
-                    new_training.rules = training.rules
-                    new_training.actions = training.actions
-                    new_training.duration = training.duration
-                    new_training.save()
-                else:
-                    print('błąd walidacji')
 
-            return redirect(mezocycles,mezocycle.team.id)
+            all_forms_valid = True
+            for form in forms_list:
+                if not form.is_valid():
+                    print("brak walidacji")
+                    all_forms_valid=False
+            if all_forms_valid:
+                new_mezocycle = Mezocycle(name=mezocycle.name, user=mezocycle.user, team=mezocycle.team, weeks=mezocycle.weeks, trainings_per_week=mezocycle.trainings_per_week)
+                mezocycle_in_db = get_object_or_404(Mezocycle,pk = mezocycle_id)
+                if new_mezocycle.name==mezocycle_in_db.name:
+                    new_mezocycle.name+="-kopia"
+                new_mezocycle.save()
+                for form in forms_list:
+                    if form.is_valid():
+                        training = form.save(commit=False)
+                        new_training = Training_in_mezocycle()
+                        new_training.mezocycle = new_mezocycle
+                        new_training.topic = training.topic
+                        new_training.week_number = training.week_number
+                        new_training.training_number = training.training_number
+                        new_training.goals = training.goals
+                        new_training.rules = training.rules
+                        new_training.actions = training.actions
+                        new_training.duration = training.duration
+                        new_training.save()
+                    else:
+                        print('błąd walidacji')
+
+                return redirect(mezocycles,mezocycle.team.id)
 
     else:  
         forms_list = []
@@ -1015,7 +1034,7 @@ def edit_mezocycle(request, mezocycle_id):
         for_t = range(1, mezocycle.trainings_per_week + 1)
         for training in trainings:
             #print(training)
-            form = Training_in_mezocycleForm(request.POST or None, instance=training, prefix=(str(training.week_number)+str(training.training_number)))
+            form = Training_in_mezocycleForm(request.POST or None, instance=training, prefix=(str(training.week_number)+'_'+str(training.training_number)))
             form.week_number = training.week_number
             form.training_number =  training.training_number
             forms_list.append(form)
@@ -1048,7 +1067,7 @@ def implement_mezocycle(request, mezocycle_id):
     print(request.POST)
     for tr in trainings_in_mezocycle:
         training = Training(topic=tr.topic,actions=tr.actions,goals=tr.goals,rules=tr.rules)
-        form = ImplementTrainingForm(request.POST or None,initial={'topic':tr.topic,'actions':tr.actions,'goals':tr.goals,'rules':tr.rules,'duration':tr.duration}, players = players,season=season, prefix=(str(tr.week_number)+str(tr.training_number)))
+        form = ImplementTrainingForm(request.POST or None,initial={'topic':tr.topic,'actions':tr.actions,'goals':tr.goals,'rules':tr.rules,'duration':tr.duration}, players = players,season=season, prefix=(str(tr.week_number)+'_'+str(tr.training_number)))
         form.week_number = tr.week_number
         form.training_number =  tr.training_number
         forms_list.append(form)
